pyproject.toml
[project]
name = "fact-bot"
version = "0.1.0"
description = "Fact bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.3",
]
